implemented SearchPosition and Compare, fixed tests

+++ Zylab.Interview.BinStorage/Index/BTree/Persistent/Sizes.cs
@@ -2,7 +2,7 @@
 
 	public class Sizes {
 		public const int Md5HashSize = 16;
-		public const int IndexDataSize = Md5HashSize + sizeof(long) + sizeof(long); // md5 hash + offset + size
+		public const int IndexDataSize = Md5HashSize + sizeof(long) + sizeof(long);
 		public const int CursorHolderSize = sizeof(long);
 		public const long RootHolderOffset = CursorHolderSize;
 		public const int RootHolderSize = sizeof(long);
@@ -11,7 +11,7 @@
 			var degree2 = degree * 2;
 
 			ChildrensSize = degree2 * sizeof(long);
-			KeysSize = (degree2 - 1) * (sizeof(long) + sizeof(long));
+			KeysSize = (degree2 - 1) * (sizeof(int) + sizeof(long)); // KeyInfo
 			NodeSize = ChildrensSize + KeysSize;
 			KeysOffset = sizeof(int) + sizeof(int); // PersistentNode.KeysCount + PersistentNode.ChildrensCount
 			ChildrensOffset = KeysSize + KeysOffset;
